move fastq_map related to collate
