+++ src/collate.hpp
@@ -11,7 +11,11 @@
  */
 #ifndef COLLATE_HPP
 #define COLLATE_HPP
+#include <htslib/sam.h>
 #include <string>
+#include "gzstream.h"
+#include "leviosam_utils.hpp"
+#include "robin_hood.h"
 
 
 struct collate_opts {
@@ -27,6 +31,14 @@ struct collate_opts {
     std::string out_r1_fname = "";
     std::string out_r2_fname = "";
 };
+
+typedef robin_hood::unordered_map<std::string, LevioSamUtils::FastqRecord> fastq_map;
+
+fastq_map read_unpaired_fq(
+    const std::string& fq_fname);
+fastq_map read_deferred_bam(
+    samFile* dsam_fp, samFile* out_dsam_fp, sam_hdr_t* hdr,
+    ogzstream& out_r1_fp, ogzstream& out_r2_fp);
 
 void print_collate_help_msg();
