+++ src/leviosam_utils.cpp
@@ -342,96 +342,6 @@ int FastqRecord::write(ogzstream& out_fq, std::string name) {
 }
 
 
-/* Read a SAM/BAM file, write properly paired alignements to a BAM file
- * and return the rest as a fastq_map
- */
-fastq_map read_deferred_bam(
-    samFile* dsam_fp, samFile* out_dsam_fp, sam_hdr_t* hdr,
-    ogzstream& out_r1_fp, ogzstream& out_r2_fp
-) {
-    fastq_map reads1, reads2;
-    bam1_t* aln = bam_init1();
-
-    while (sam_read1(dsam_fp, hdr, aln) > 0) {
-        bam1_core_t c = aln->core;
-        // The following categories of reads are excluded by this method
-        if ((c.flag & BAM_FSECONDARY) || // Secondary alignment - no SEQ field
-            (c.flag & BAM_FQCFAIL) || // not passing filters
-            (c.flag & BAM_FDUP) || // PCR or optinal duplicate
-            (c.flag & BAM_FSUPPLEMENTARY)) { // supplementary alignment
-            continue;
-        }
-        std::string qname = bam_get_qname(aln);
-        FastqRecord fq = FastqRecord(aln);
-        if (c.flag & BAM_FREAD1) { // first segment
-            auto search = reads2.find(qname);
-            if (search != reads2.end()) {
-                if (sam_write1(out_dsam_fp, hdr, search->second.aln) < 0  ||
-                    sam_write1(out_dsam_fp, hdr, aln) < 0) {
-                    std::cerr << "[Error] Failed to write record " << qname << "\n";
-                    exit(1);
-                }
-                fq.write(out_r1_fp, qname);
-                search->second.write(out_r2_fp, qname);
-                reads2.erase(search);
-            } else {
-                reads1.emplace(std::make_pair(qname, fq));
-            }
-        } else if (c.flag & BAM_FREAD2) { // second segment
-            auto search = reads1.find(qname);
-            if (search != reads1.end()) {
-                if (sam_write1(out_dsam_fp, hdr, aln) < 0 ||
-                    sam_write1(out_dsam_fp, hdr, search->second.aln) < 0) {
-                    std::cerr << "[Error] Failed to write record " << qname << "\n";
-                    exit(1);
-                }
-                search->second.write(out_r1_fp, qname);
-                fq.write(out_r2_fp, qname);
-                reads1.erase(search);
-            } else {
-                reads2.emplace(std::make_pair(qname, fq));
-            }
-        } else {
-            std::cerr << "[Error] Alignment " << qname << " is not paired.\n";
-            exit(1);
-        }
-    }
-    std::cerr << "Num. unpaired reads1: " << reads1.size() << "\n";
-    std::cerr << "Num. unpaired reads2: " << reads2.size() << "\n";
-    for (auto& r: reads2) {
-        reads1.emplace(std::make_pair(r.first, r.second.aln));
-    }
-    std::cerr << "Num. merged unpaired: " << reads1.size() << "\n";
-    return reads1;
-}
-
-
-/* Read a single-end FASTQ file and return a fastq_map */
-fastq_map read_unpaired_fq(const std::string& fq_fname) {
-    fastq_map reads;
-    std::ifstream fastq_fp(fq_fname);
-    std::string line;
-    int i = 0;
-    std::string name, seq;
-    while (getline (fastq_fp, line)) {
-        if (i % 4 == 0) {
-            name = line.substr(1);
-        } else if (i % 4 == 1) {
-            seq = line;
-        } else if (i % 4 == 3) {
-            reads.emplace(
-                std::make_pair(name, FastqRecord(seq, line)));
-            name = "";
-            seq = "";
-        }
-        i++;
-    }
-    std::cerr << "Number of singletons: " << reads.size() << "\n";
-    fastq_fp.close();
-    return reads;
-}
-
-
 /* Split a string on a delimiter
  * From: https://stackoverflow.com/a/64886763
  */
