+++ src/leviosam_utils.hpp
@@ -20,7 +20,6 @@
 #include <cstring>
 #include "bed.hpp"
 #include "gzstream.h"
-#include "robin_hood.h"
 
 const int8_t seq_comp_table[16] = { 0, 8, 4, 12, 2, 10, 6, 14, 1, 9, 5, 13, 3, 11, 7, 15 };
 
@@ -64,14 +63,6 @@ public:
     std::string qual_str;
     bam1_t* aln = NULL;
 };
-
-typedef robin_hood::unordered_map<std::string, FastqRecord> fastq_map;
-
-fastq_map read_unpaired_fq(
-    const std::string& fq_fname);
-fastq_map read_deferred_bam(
-    samFile* dsam_fp, samFile* out_dsam_fp, sam_hdr_t* hdr,
-    ogzstream& out_r1_fp, ogzstream& out_r2_fp);
 
 class WriteDeferred {
 public:
